requirements.txt
torch
torchvision
matplotlib
pandas
optuna
datasets
Pillow
tqdm
